Version 1.0

+++ server/controllers/seguridad/seguridad_perfiles_menu_submenu.controller.js
@@ -1,9 +1,25 @@
+/* eslint-disable no-throw-literal */
 const SeguridadPerfilesMenuSubmenu = require("../../models/seguridad/seguridad_perfiles_menu_submenu.model");
 
-const index = (req, res) => {
+/**
+ * 
+ * @param {Object} request
+ * @param {Object} response
+ * @returns Lista todas las subpaginas de Perfil con su respectivo Menu y Submenu
+ */
+
+const index = async (req, res) => {
   try {
     const { id } = req.params
-    const menusAndSubmenus = SeguridadPerfilesMenuSubmenu.findOne({ ID_SEGURIDAD_PERFILES: id }).populate("ID_CONFIGURACION_MENU").populate("ID_CONFIGURACION_SUBMENU")
+
+    if (!id) throw({
+      error: true,
+      status: 404,
+      statusText: "Parametro Requerido no encontrado"
+    })
+
+    const menusAndSubmenus = await SeguridadPerfilesMenuSubmenu.findOne({ ID_SEGURIDAD_PERFILES: id }).populate("ID_CONFIGURACION_MENU").populate("ID_CONFIGURACION_SUBMENU")
+
     res.status(201).json({
       error: false,
       status: 201,
@@ -15,9 +31,34 @@ const index = (req, res) => {
   }
 }
 
+/**
+ * 
+ * @param {Object} request
+ * @param {Object} response
+ * @returns Guarda Sub Paginas por Perfil de Usuario
+ */
+
 const store = (req, res) => {
   try {
     const { PAGINAS_PERFIL, ID_PERFIL } = req.body;
+    
+    if (PAGINAS_PERFIL instanceof Array === false) throw({
+      error: true,
+      status: 401,
+      statusText: "Parametro PAGINAS_PERFIL no es un arreglo"
+    })
+
+    if (PAGINAS_PERFIL.length === 0) throw({
+      error: true,
+      status: 401,
+      statusText: "Parametro PAGINAS_PERFIL esta vacio"
+    })
+
+    if (!ID_PERFIL) throw({
+      error: true,
+      status: 401,
+      statusText: "Parametro ID_PERFIL es requerido"
+    })
 
     Promise.all([
       SeguridadPerfilesMenuSubmenu.deleteMany({ ID_SEGURIDAD_PERFILES: ID_PERFIL })
@@ -36,6 +77,12 @@ const store = (req, res) => {
   }
 }
 
+/**
+ * 
+ * @param {Object} request
+ * @param {Object} response
+ * @returns Filtra las sub paginas por perfil de usuario
+ */
 
 const show = async (req, res) => {
   try {
